Move template to separate file to fix nawk bug
Setting multi-line variables from the shell script results in an error
when using nawk as the awk implementation.  Since we're aiming for
portability, this is a big no-no for us.

Unfortunately, the only other way we know for setting templates in awk
is by importing them from within the awk script itself.  Not a fan of
having a completely separate file for this lone purpose, but it can't be
helped.  If anyone reading this knows of an alternate method that works
on gawk, mawk *and* nawk, we're all ears.